ucl_api/requirements.txt
Django==4.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-spectacular==0.27.2
gunicorn==23.0.0
pillow==9.0.0
psycopg2==2.9.3
uuid==1.30
